payment-service/requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.8
django-cors-headers==3.10.0
django-environ==0.8.1
django-phonenumber-field==5.2.0
djangogrpcframework==0.2.1
djangorestframework==3.12.4
grpcio==1.41.1
grpcio-tools==1.41.1
idna==3.3
phonenumbers==8.12.35
pika==1.2.0
pika-stubs==0.1.3
protobuf==3.19.0
psycopg2-binary==2.9.1
pycodestyle==2.8.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
stripe==2.61.0
toml==0.10.2
urllib3==1.26.7
